NCM - khurt blood

+++ project/Game/Source/Enemy_Boorok.cpp
@@ -11,6 +11,7 @@
 #include "Physics.h"
 #include "Window.h"
 #include "Pathfinding.h"
+#include "ParticleSystem.h"
 #include "Map.h"
 #include "Physics.h"
 #include "Item_Hueso.h"
@@ -278,7 +279,7 @@ void Enemy_Boorok::Die()
 	pbodyFoot->body->SetLinearVelocity(b2Vec2_zero);
 	currentAnimation = &dieAnim;
 
-	fPoint pos(position.x, position.y);
+	fPoint pos((float)position.x, (float)position.y);
 	blood = app->psystem->AddEmiter(pos, EMITTER_TYPE_ENEMY_BLOOD);
 
 	if (dieAnim.HasFinished())
